fix empty day warnings
