+++ src/redux/selectors.js
@@ -8,6 +8,10 @@ import { selectAllWeekEntriesByActivityId } from './LogSlice'
 
 import { getTodayTime, startOfDay } from './../util'
 
+import { dueToday, dueThisWeek } from '../activityHandler'
+
+import { selectEntryByActivityIdAndDate } from './LogSlice'
+
 /* 
   This file defines selectors that use data from more than one slice 
   and/or is out of the slice responsibilities
@@ -91,4 +95,22 @@ export function selectAllActivitiesByGoalId(state, goalId){
   })
   
   return thisGoalActivities 
+}
+
+export function selectVisibleActivities(state, date){
+  const activities = selectAllActiveActivitiesByDate(state, date)
+
+  const visibleActivities = activities.filter(a => {
+    if(dueToday(state, a.id, date)){
+      return true
+    }
+    if(dueThisWeek(state, a.id, date)){
+      const entry = selectEntryByActivityIdAndDate(state, a.id, date)
+      if(!entry.archived){
+        return true
+      }
+    }
+    return false
+  })
+  return visibleActivities
 }
